Flights.ai/requirements.txt
gradio
requests
gtts
ollama
python-dotenv
